os1sem/2021_10_22_OS/HomeWork_part2/KKCO-03/KomogorovKirill/requirement.txt
fuse-python
fastapi
